Moved Config.jsx inside src and fixed imports
