+++ src/components/FarmerDashboard.jsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { api } from "../../Config";
-// import { api } from "/home/zakarie/agrolink/Config.jsx";
+import { api } from "../Config";
 import "./DashboardsAndProfile.css";
 
 const FarmerDashboard = () => {
@@ -52,3 +51,4 @@ const FarmerDashboard = () => {
 };
 
 export default FarmerDashboard;
+
