+++ src/components/Login.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import "./Login.css";
-import { api } from "../../Config";
+import config from "../Config.jsx";
 import { useNavigate } from "react-router-dom";
 // import { api } from "https://agrolink-backend-nezp.onrender.com";
